Zoop.MarketPlace.Buyers (Zoop v0.1.0)

Essa classe é resposavel por lidar com os usuarios dentro do marketplace ao nivel do marketplace zoop.

Functions

create_buyer(buyers)

Cria o usuario dentro do markeplace ('não é associado ao vendedor')

delete_buyer(user_id)

Deleta um usuario do marketplace utilizando como parametro

get_all_buyers()

Lista todos os usuarios do marketplace ('não realiza associação com o vendedor')

get_buyer(user_id)

Pega os dados do usuario associado ao id passado como parametro.

search_buyer(taxpayer_id)

Busca os usuario por CPF/CNPJ

update_buyer(user_id, buyers)

Atualiza os dados do comprador